Drivers: Intel: DMIC: Add NHLT binary data based configuration
This patch adds to DMIC driver support for binary data based HW
registers setup. The driver mode is selected in Kconfig.

This WIP version carries the needed part of the blob in the header
file nhlt_dmic_cfg.h. It needs to be replaced by an IPC passed blob.
The default hard coded blob can work only in 38.4 MHz rate IOclock
devices. There are other configurations to select via the defines in
the header file.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>